% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tweet.R
\name{action_auto_tweet}
\alias{action_auto_tweet}
\title{Post SER tweets from the tweet queue Monday through Friday}
\usage{
action_auto_tweet(twitter_token = ser_token,
  google_drive_auth = drive_auth_token())
}
\arguments{
\item{twitter_token}{an rtweet token for the SER app}

\item{google_drive_auth}{the name of the google drive authorization file}
}
\value{
a data frame containing the updated tweet queue
}
\description{
\code{action_auto_tweet()} handles daily tweeting on the SER account. It randomly pulls
one tweet from the tweet queue per day to post. It cycles through each tweet
before posting a tweet again. Additionally, there is a 1 in 10 chance that it
will instead retweet a popular SER tweet. This function will not post
anything if it is a blackout date. It causes several side effects, including
downloading the tweet and retweet queue, posting tweets, and updating the
files on Google Drive. Requires Google Drive and Twitter authorization.
}
