% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_dashboard.R
\name{action_deploy_dashboard}
\alias{action_deploy_dashboard}
\alias{create_dashboard}
\title{Deploy the SER Tweet Dashboard to shinyapps.io}
\usage{
action_deploy_dashboard(path = ".", create = TRUE)

create_dashboard(path = ".")
}
\arguments{
\item{path}{The path where the dashboard exists or should be created}

\item{create}{Create the dashboard?}
}
\value{
Invisibly, \code{path}
}
\description{
\code{action_deploy_dashboard()} deploys the Twitter dashboard to \code{shinyapps.io}.
Use \code{create_dashboard()} to make a copy of the dashboard and download the
data without deploying.
}
