% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_summary.R
\name{action_email_summary}
\alias{action_email_summary}
\title{Email Daily Summary of SER Tweets}
\usage{
action_email_summary(recipients, twitter_token = ser_token)
}
\arguments{
\item{recipients}{a vector of email addresses to blind copy the email to}

\item{twitter_token}{an rtweet token for the SER app}

\item{gmail_id, gmail_secret}{gmail authorization}
}
\value{
the email to be sent, a character vector
}
\description{
\code{action_email_summary()} handles daily summaries of tweets made on the SER
account. This function emails an HTML table (using the gt package) of the
tweets and retweets for the day to a list of recipients. It returns a
character vector containing the email message that it sent. Requires Gmail
and Twitter authorization.
}
