% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_meeting_tweets.R
\name{action_meeting_tweet}
\alias{action_meeting_tweet}
\title{Post tweets about the SER Annual Meeting}
\usage{
action_meeting_tweet(
  period_start = "2019-03-01",
  period_end = "2019-06-18",
  twitter_token = ser_token,
  google_drive_auth = "ttt.rds"
)
}
\arguments{
\item{period_start}{beginning of period to tweet about meeting}

\item{period_end}{end of period to tweet about meeting}

\item{twitter_token}{an rtweet token for the SER app}

\item{google_drive_auth}{the name of the google drive authorization file}
}
\value{
a data frame containing the updated meeting tweet queue
}
\description{
\code{action_meeting_tweet()} handles yearly tweeting about the SER Annual
Meeting. It randomly pulls one tweet from the meeting tweet queue to post.
The function will run March to June, posting 3 tweets a week until May the 5
tweets a week until the meeting starts. This function will not post anything
if it is a blackout date. It causes several side effects, including
downloading the meeting tweet queue and posting tweets. Requires Google Drive
and Twitter authorization.
}
