% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_ghactions.R
\name{deploy_ghactions}
\alias{deploy_ghactions}
\title{Prepare a repository to deploy an SER script on GitHub Actions}
\usage{
deploy_ghactions(script_name, cron = "30 16 * * *")
}
\arguments{
\item{script_name}{The name of the script, which should be included in
\code{inst/scripts}}

\item{cron}{A cron job specification in include in the GitHub Action script}
}
\description{
Use this function after setting up a repository to run the script in. Add
environmental variables under Settings > secrets in the GitHub UI for the
repository.
}
