% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_data.R
\name{download_twitter_data}
\alias{download_twitter_data}
\alias{upload_twitter_data}
\alias{update_twitter_data}
\alias{action_collect_twitter_data}
\title{Download, update, and upload SER twitter data}
\usage{
download_twitter_data(path = ".", google_drive_auth = drive_auth_token())

upload_twitter_data(path = ".", google_drive_auth = drive_auth_token())

update_twitter_data(path = ".", google_drive_auth = drive_auth_token())

action_collect_twitter_data(path = ".", google_drive_auth = drive_auth_token())
}
\arguments{
\item{path}{Path to data}

\item{google_drive_auth}{A Google Drive token}
}
\value{
invisibly, \code{path}
}
\description{
\code{action_collect_twitter_data()} handles Twitter data collection related to
the SER account, primarily used in \code{\link[=action_deploy_dashboard]{action_deploy_dashboard()}}. It
downloads, updates, and uploads the data, which is stored on Google Drive.
}
