% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tweet.R
\name{drive_auth_token}
\alias{drive_auth_token}
\alias{ser_token}
\title{Get tokens for auto-tweet functionality}
\usage{
drive_auth_token()

ser_token()
}
\value{
A Twitter or Google Drive token
}
\description{
Get tokens for auto-tweet functionality
}
