% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{email}
\alias{email}
\alias{gmail}
\title{Compose email addresses}
\usage{
email(x, y)

gmail(x)
}
\arguments{
\item{x}{the usernname}

\item{y}{the domain name}
}
\description{
A set of functions to compose email addresses in public scripts to avoid
scraping by bots. \code{gmail()} is a wrapper around \code{email()} that always appends
the gmail domain.
}
