% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_email.R
\name{on_error_email_to}
\alias{on_error_email_to}
\alias{email_to}
\alias{gmail_id}
\alias{gmail_secret}
\alias{email_on_error}
\alias{action_safely}
\title{Email errors in code}
\usage{
on_error_email_to(recipient, gmail_id, gmail_secret)

email_to()

gmail_id()

gmail_secret()

email_on_error(.e, recipient = email_to(), path)

action_safely(.f)
}
\arguments{
\item{recipient}{an email address.}

\item{.e}{the error message}

\item{gmail_secret_file}{the name of the gmail authorization file}
}
\value{
a character vector containing the error email
}
\description{
\code{email_on_error()} will email on error. Use it with \code{action_safely()} to wrap
a function using the emailer. Set the email recipient globally with
\code{on_error_email_to()} and retrieve it with \code{email_to()}.
}
