% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_summary.R
\name{get_content_lead}
\alias{get_content_lead}
\title{Retrieve Email for Active Content Lead}
\usage{
get_content_lead(google_drive_auth = drive_auth_token())
}
\arguments{
\item{google_drive_auth}{Authorization token for the SER Google Drive account.}
}
\description{
\code{get_content_lead()} retrieves the email of teh currently assigned content
lead. The function retrieves a list of content leads and their assignment
dates, checks these against the current date, and returns a character string
containing an email address.
}
