% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro-data.R
\docType{data}
\name{hydro_df}
\alias{hydro_df}
\title{Discharge data in a stream}
\format{
A dataframe
   \describe{
      \item{Date}{Date data, from 2016 to 2018}
      \item{Discharge}{Discharge data (L/s), daily discharge data in a stream}
   }
}
\source{

}
\usage{
hydro_df
}
\description{
Discharge data in a stream
}
\keyword{datasets}
