% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_date-data.R
\docType{data}
\name{sample_date}
\alias{sample_date}
\title{Sample dates (vector)}
\format{
A vector containg 13 dates data
}
\usage{
sample_date
}
\description{
Sample dates (vector)
}
\keyword{datasets}
