% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SER.R
\name{SER}
\alias{SER}
\title{Short-period Environmental Regime (SER)}
\usage{
SER(Env_data, sample_date, days_bf = NULL)
}
\arguments{
\item{Env_data:}{time series environmental data, a dataframe with columns Date and Env, make sure Date is in date format while Env is in numeric format.}

\item{sample_date:}{sampling dates, a vector containing sampling days. Date for initiation of the experiment (measurements) should be included as the first element of the vector.}

\item{days_bf:}{N days before the sampling date (not included); by default, days_bf = c(days between two successive sampling dates)}
}
\value{
a data frame
}
\description{
Short-period Environmental Regime (SER)
}
\examples{
# inspect the discharge data
str(hydro_df)

# make sure name of the column with data (in this case Discharge) is "Env"
names(hydro_df)[2] <- "Env"

# inspect the sample date data
str(sample_date)

# calculate short-period hydrological indices
hydro_ser <- SER(hydro_df,sample_date,days_bf = NULL)
str(hydro_ser)
head(hydro_ser)


}
