% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SER.R
\name{SERs}
\alias{SERs}
\title{Short-period environmental regime for multiple environmental measurements}
\usage{
SERs(
  df.ls,
  date.ls,
  days_bf = NULL,
  type = NULL,
  include_sample.date = FALSE,
  include_successive = FALSE,
  simiplify = TRUE
)
}
\arguments{
\item{df.ls}{a list containing several dataframes. Please see the help of SER for the requirements of each dataframe in the list.}

\item{date.ls}{a list containing sampling dates for each dataframe in df.ls.}

\item{days_bf}{see the help of SER.}

\item{type}{see the help of SER.}

\item{include_sample.date}{see the help of SER.}

\item{include_successive}{see the help of SER.}

\item{simiplify}{see the help of SER.}
}
\value{
a list containing SER for each dataframes and corresponding sampling dates.
}
\description{
Short-period environmental regime for multiple environmental measurements
}
\examples{
# generate a df.ls
names(hydro_df)[2] <- "Env"
df.ls <- list(env1 = hydro_df, env2 = hydro_df[sample(1:nrow(hydro_df), nrow(hydro_df), replace = FALSE), ])

# generate a date.ls
date.ls <- list(dt1 = sample_date, dt2 = sample_date[1:6])

# calculate SER
(SERs(df.ls, date.ls))
}
