% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_ser.R
\name{source_ser}
\alias{source_ser}
\alias{use_ser_script}
\alias{scripts_ls}
\title{Create and run SER scripts}
\usage{
source_ser(name)

use_ser_script(name)

scripts_ls(name)
}
\arguments{
\item{name}{The name of the script (without the \code{.R} extension)}
}
\description{
\code{source_ser()} runs the available scripts in the ser package.
\code{use_ser_script()} sets up a new script to set deply, which is added to
\code{inst/scripts} in the package. \code{scripts_ls()} lists the currently available
scripts.
}
