% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{drive_auth_token}
\alias{drive_auth_token}
\alias{ser_token}
\alias{sandbox_token}
\alias{rsconnect_auth}
\title{Get tokens for SER functionality}
\usage{
drive_auth_token()

ser_token()

sandbox_token()

rsconnect_auth()
}
\value{
A Twitter or Google Drive token
}
\description{
Get tokens for SER functionality
}
