% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SER.R
\name{SER}
\alias{SER}
\title{Short-period Environmental Regime (SER)}
\usage{
SER(
  Env_data,
  sample_date,
  days_bf = NULL,
  type = NULL,
  include_sample.date = FALSE,
  include_successive = FALSE,
  simiplify = TRUE
)
}
\arguments{
\item{Env_data:}{time series environmental data, a dataframe with columns Date and Env, make sure Date is in date format while Env is in numeric format.}

\item{sample_date:}{sampling dates, a vector containing sampling days. Date for initiation of the experiment (measurements) should be included as the first element of the vector.}

\item{days_bf:}{a numeric vector representing N days (months or years) before the sampling date; if days_bf = NULL (by default), then it days_bf = days between two successive sampling dates}

\item{type:}{a character vector indicating whether the 'days_bf' is "day", "month", or "year" specific.}

\item{include_sample.date:}{TRUE or FALSE (by default) indicates whether environmental data at sample.date is included during the calculation.}

\item{include_successive:}{TRUE or FALS (by default) indicates whether environmental data between two successive sampling dates is used for calculating 'BetwSamT' environmental regime.}

\item{simiplify:}{TRUE (by default) or FALSE indicates whether the simplified result is returned. Set simplify = TRUE if you want to check the intermediate results.}
}
\value{
a data frame
}
\description{
Short-period Environmental Regime (SER)
}
\examples{
# inspect the discharge data
str(hydro_df)

# make sure name of the column with data (in this case Discharge) is "Env"
names(hydro_df)[2] <- "Env"

# inspect the sample date data
str(sample_date)

#############
#day- specific SER
# calculate short-period hydrological indices
SER(hydro_df,sample_date)
SER(hydro_df,sample_date, days_bf=c(3,7,14))
SER(hydro_df,sample_date, days_bf=c(3,7,14), include_sample.date = T)
SER(hydro_df,sample_date, days_bf=c(3,7,14), include_successive = T)
SER(hydro_df,sample_date, days_bf=c(3,7,14), include_sample.date = T, include_successive = T)

##################
# month-specific SER
# only include the first and last dates in sample_dates
test.date.m <- sample_date[c(1, 13)]
SER(hydro_df, test.date.m, days_bf = NULL, type = "month")
SER(hydro_df, test.date.m, days_bf = c(2, 4, 6), type = "month", include_sample.date = T)
SER(hydro_df, test.date.m, days_bf = c(2, 4, 6), type = "month", include_successive = T)
SER(hydro_df, test.date.m, days_bf = c(2, 4, 6), type = "month", include_sample.date = T, include_successive = T)

##################
# year-specific SER
# generate data for testing
test.df <- data.frame(Date = seq(ymd("2010-1-1"), ymd("2020-1-1"), by="1 day"),
                      Env = sample(hydro_df$Env, 3653, replace = T))
test.date.y <-  seq(ymd("2012-1-1"), ymd("2020-1-1"), by="2 year")
SER(test.df, test.date.y, days_bf = NULL, type = "year")
SER(test.df, test.date.y, days_bf = c(2), type = "year", include_sample.date = T)
SER(test.df, test.date.y, days_bf = c(2), type = "year", include_successive = T)
SER(test.df, test.date.y, days_bf = c(2), type = "year", include_successive = T, include_sample.date = T)
}
